Se agregaron lineas de codigo que permiten modificaciones al modelo

+++ Emotion.Web/Models/EmoEmotion.cs
@@ -8,7 +8,7 @@ namespace Emotion.Web.Models
     public class EmoEmotion
     {
         public int Id { get; set; }
-        public float Score { get; set; }
+        public Decimal Score { get; set; }
         public int EmoFaceId { get; set; }
         public EmoEmotionEnum EmotionType{ get; set; }
